pyproject.toml
[project]
name = "competitor_check"
version = "0.0.0"
description = "Validates that LLM-generated text is not naming any competitors from a given list."
authors = [
    {name = "Guardrails AI", email = "contact@guardrailsai.com"}
]
license = {file = "LICENSE"}
readme = "README.md"
requires-python = ">= 3.9"
dependencies = [
    "guardrails-ai>=0.6.0",
    "sentence-splitter>=1.4",
    "transformers>=4.40.2",
    "torch>=2.0"
]

[project.optional-dependencies]
dev = [
    "pyright",
    "pytest",
    "ruff"
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-rP"
testpaths = [
    "test"
]

[tool.pyright]
include = ["validator"]
